pyproject.toml
[project]
name = "demo-streaamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.4.0",
    "google-generativeai>=0.8.5",
    "pandas>=2.3.2",
    "streamlit>=1.50.0",
]
